the application checks more thoroughly if it can go in run mode: it now checks presence and protocol version of all attached can boards (but not mtb yet).
in more details.

at bootstrap, the ems starts checking the presence of all the can boards it has (except for mtb which does not have the get-fw-version message yet).
the discovery is done in chain by board type and with a timeout.
for instance, in eb1 the EOCANdiscovery2 searches all the 2foc boards with a 3 seconds timeout, then if the previous search is ok, it searches
the strain with 3 seconds timeout.

at the end of the discovery chain the application set a flag which allows to go in run mode.

the ems accepts a goto-run command only if the previous flag is true, i.e., when all the expected can boards are present and all have
the correct protocol version.

if the ems is asked by robotinterface to go in run mode but it cannot, the ems sends up to the last report of EOtheCANdiscovery2 which
causes the failure. in such a way the user knows which boards are not present.

so far what to check is still in ROM of the ems board and not yet sent by robotinterface.