refactoring of can parsing.
now we have some can parsers, each one dedicated to a specific set of can messages which expose a pure virtual class
with the methods that an agent must have in order to operate on messages. and each agent must be derived from that
class. the can parsers at init() accept a pointer to the agent and call it at reception of a can message.

example:
the object theCANparserIMU exposes a CANagentIMU interface and at reception of an IMU message calls the proper
method of the interface.
the object theIMU is now derived from CANagentIMU and must implement its interfaces so that the handling of
the CAn message is its responsibility.
the object theCANparserIMU receives an handle of theIMU (but possibly of any other object derived from CANagentIMU)
so that it delegates to it the action of a received can message.
we needed that because the can parser as implemented before were getting too complicated and dedicated to a specific
agent.