rfe: new version 1.2.0. solves a possible freeze of the application (likely to happen with bursts of usb packets for facial expressions)
here are details:
the application could freeze in a forever loop called by the main thread which process every event (can rx frames, usb rex frames,
IMU tx requests etc). the result was that the board was unreachable until a restart was executed.
The forever loop was never exited because for some unknown reason the HAL_SPI_TxCpltCallback() which STM HAL calls at end of
SPI tx ... is not called.

i have added some code which prevents such a freeze.
- the spi tx buffer containing the LED info is now never modified during an SPI DMA transmission.
- the code which waits for the end of an SPI DMA transmission has now a timeout so that is the HAL_SPI_TxCpltCallback() is never
  called for whatever unknown reason, the main thread can safely continue.
- there is code (so far non activated) which sends a CAN print in case of a timeout episode.

i have done some stress tests which have caused the timeout and have shown a safe recovery.

as a further safety measure, i have also increased the stack size of the main thread, situation for which we did not experience any timeout.