osal -> enhanced safety behavior before initialization of osal ...
sometimes, it may happen by mistake that someone calls a osal_xxx_new() or any other osal_*() function before
the initialization of the RTOS.

yes, it is wrong but it may happen.

case 1.
for instance if static objects have constructors with memory allocation or even worse with creation of rtos objects ...
the above constructors are called just before the main() in a moment when the user defined error handler of osal
is not initted yet. in this case the default error handler comes to help. BUT that enters in a forever loop
inside the osal.lib and teh debugger never enters main().... very nasty behavior.

case 2.
or maybe someone just wrongly calls a osal_xxx_new() between start of main() and init of the RTOS.

for this reason in here:

- we add a function which can set the error handler: osal_set_errorhandler()
  the use of it is inside embot::os::rtos and sets a user-defined error handler guaranteed
  to be ready at start of main(). that catches the second case.

- we add code in osal so that every osal_xxx_new() or other osal_*() calls detect the three cases
  (a) call from an ISR, (b) call when osal has started, (c) call when osal has not started yet
  and behave accordingly.