hal_spiencoder: fixed a memory leak
multiple cycles of {init(), deinit()} of the spiencoder wasted heap because the cleaning phase of deinit() was
incomplete. now we either clean everything (if macro SPIENC_DEINIT_DEALLOCATE_HEAP is defined) or we try to
keep the memory in order to avoid delete / new situations in runtime. (this latter is the default mode).