Add fix for not causing joints to go in hw fault: limits exceeded at yarprobotinteface restarting
Fault visible only on MAIS due to the restoring of default values at
yarprobotinterface restart.
This is just a temporary fix, which allows to work more easily on iCub3
robots and it will be substituted by a proper patch soon.
Update versioning for fw and fix error for changed method in amc