2FOC current filter implemented.
Since PWM and current feedback are transmitted at 1 kHz, but set/measured at 20 kHz, I've introduced a filter acting in a 20 samples wide window.
The PWM 1 kHz feedback is the average of 20 values in a 1 msec window.
The current 1 kHz feedback is the average of 14 samples in a 1 msec window because the 3 highest and the 3 lowest values are discarded.
Moreover, the current signal for current control is the average of two consecutive samples.
The ADC is calibrated at startup (offset and gain).
I've removed the assembly routines for current calibration because anachronistic and difficult to maintain, and replaced with two lines of C code.